meta-lxatac-software: bundles: hook.sh: enable the cert signing this bundle
This makes it so that an image installed via rauc will trust only the key the
bundle was signed with.

This will be used to support our official bundle release channels, where
bundles will be re-signed multiple times in the release process.

First the bundles are built and signed with the rauc-devel.key.pem key and
tested on hardware.
If the tests are passed the existing bundle will be automatically re-signed
with the nightly release key.
If a new stable release is to be released an existing nightly bundle that has
been manually tested is manually re-signed with the stable release key.

All of these bundles are deployed with the same /etc/rauc/certificates-*
but a stable bundle should only be able to install other stable bundles,
so we need to detect the release channel at install-time.

This commit makes it so that only the certificate matching the key the bundle
was signed with is enabled.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>